refactor: remove unused components and hooks from auth and navigation features
- Deleted `useButtonTriggerState` hook and related props from LoginButtonTrigger.
- Removed `LoginFormContent` placeholder component.
- Eliminated `usePopoverState` hook and associated styles from LoginPopoverWrapper.
- Cleaned up `MainMenu` component and its related hooks, constants, and types.
- Introduced `MainMenuDesktop` and `MainMenuMobile` components for better separation of mobile and desktop navigation.
- Updated `Header` component to utilize the new `MainMenuDesktop` and `MainMenuMobile`.
- Refactored `Pricing` component to directly define pricing plans instead of importing from a separate file.
- Added `LanguageSelect` component for language switching functionality.
