+++ src/apps/web/src/features/auth/components/LoginButton/components/LoginPopoverWrapper/LoginPopoverWrapper.tsx
@@ -1,15 +1,25 @@
 "use client";
 
-import { memo } from "react";
+import { memo, useCallback } from "react";
 
 import { Popover, PopoverContent } from "@heroui/react";
 
 import { LoginButtonTrigger } from "../LoginButtonTrigger/LoginButtonTrigger";
-import { LoginFormContent } from "../LoginFormContent";
-import { usePopoverState } from "./hooks/usePopoverState";
-import { loginPopoverWrapperStyles } from "./styles/loginPopoverWrapperStyles";
 
-import type { LoginPopoverWrapperProps } from "./types/LoginPopoverWrapperProps";
+import type { RefObject } from "react";
+
+interface LoginPopoverWrapperProps {
+  scrolled?: boolean;
+  loginFlyoutOpen: boolean;
+  setLoginFlyoutOpen: (open: boolean) => void;
+  setLoginFlyoutOpenedByHover: (opened: boolean) => void;
+  loginTriggerRef: RefObject<HTMLButtonElement | null>;
+  isHovered: boolean;
+  handleMouseEnter: () => void;
+  handleMouseLeave: () => void;
+  handleFocusIn: () => void;
+  handleFocusOut: () => void;
+}
 
 /**
  * LoginPopoverWrapper - Popover container with trigger and content
@@ -27,13 +37,14 @@ export const LoginPopoverWrapper = memo(
     handleFocusIn,
     handleFocusOut,
   }: LoginPopoverWrapperProps) => {
-    // Get styles from tailwind-variants
-    const styles = loginPopoverWrapperStyles();
-
-    // PERFORMANCE: Extract popover state logic to custom hook
-    const { handleOpenChange } = usePopoverState(
-      setLoginFlyoutOpen,
-      setLoginFlyoutOpenedByHover,
+    const handleOpenChange = useCallback(
+      (isOpen: boolean) => {
+        setLoginFlyoutOpen(isOpen);
+        if (isOpen) {
+          setLoginFlyoutOpenedByHover(false);
+        }
+      },
+      [setLoginFlyoutOpen, setLoginFlyoutOpenedByHover],
     );
 
     return (
@@ -46,16 +57,23 @@ export const LoginPopoverWrapper = memo(
           isHovered={isHovered}
         />
 
-        <PopoverContent
-          className={`${styles.popoverContent()} ${styles.content()}`}
-        >
+        <PopoverContent className="z-50 w-80 p-0">
           <div
             onMouseEnter={handleMouseEnter}
             onMouseLeave={handleMouseLeave}
             onFocus={handleFocusIn}
             onBlur={handleFocusOut}
           >
-            <LoginFormContent />
+            <div className="relative overflow-hidden rounded-xl animate-in fade-in-0 zoom-in-95 duration-200">
+              <div className="relative backdrop-blur-md bg-white/85 dark:bg-slate-900/85 border border-white/40 dark:border-slate-700/60 shadow-lg">
+                <div className="flex flex-col items-center justify-center py-12 px-8 text-center">
+                  <div className="w-16 h-16 mb-6 rounded-full bg-linear-to-br from-purple-400/30 to-blue-400/30 dark:from-purple-500/20 dark:to-blue-500/20 animate-pulse" />
+                  <p className="text-sm text-gray-500 dark:text-gray-400">
+                    Přihlášení se připravuje...
+                  </p>
+                </div>
+              </div>
+            </div>
           </div>
         </PopoverContent>
       </Popover>
