--- src/apps/web/src/features/auth/components/LoginButton/constants/hoverTiming.ts
@@ -1,7 +1,0 @@
-/**
- * HOVER_TIMING - Timing constants for hover behavior
- */
-export const HOVER_TIMING = {
-  OPEN_DELAY: 150,
-  CLOSE_DELAY: 1000,
-} as const;
