--- src/apps/web/src/features/locale/components/LanguageSelect/index.ts
@@ -1,1 +1,0 @@
-export { LanguageSelect } from "./LanguageSelect";
