+++ src/apps/web/src/features/locale/index.ts
@@ -1,4 +1,4 @@
-export { LanguageSelect } from "./components/LanguageSelect/LanguageSelect";
+export { LanguageSelect } from "./components/LanguageSelect";
 export { LanguageEnum, LanguageEnumKeys } from "./constants/LanguageEnum";
 export type {
   LanguageEnumKey,
