--- src/apps/web/src/features/navigation/components/MainMenu/constants/NavLinks.ts
@@ -1,5 +1,0 @@
-export const NAV_LINKS = [
-  { href: "#features", label: "Vlastnosti" },
-  { href: "#motivation", label: "Proč my?" },
-  { href: "#pricing", label: "Ceník" },
-];
