--- src/apps/web/src/features/navigation/components/MainMenu/hooks/useMainMenuSpotlight.ts
@@ -1,41 +1,0 @@
-import { useState } from "react";
-
-import type { MainMenuSpotlight } from "../types/MainMenuSpotlight";
-import type { MouseEvent, RefObject } from "react";
-
-export const useMainMenuSpotlight = (
-  containerRef: RefObject<HTMLDivElement | null>,
-) => {
-  const [hoveredIdx, setHoveredIdx] = useState<number | null>(null);
-  const [spotlightPos, setSpotlightPos] = useState<MainMenuSpotlight>({
-    left: 0,
-    width: 0,
-    opacity: 0,
-  });
-
-  const handleMouseEnter = (idx: number, e: MouseEvent<HTMLAnchorElement>) => {
-    const rect = e.currentTarget.getBoundingClientRect();
-    const containerRect = containerRef.current?.getBoundingClientRect();
-
-    if (containerRect) {
-      setHoveredIdx(idx);
-      setSpotlightPos({
-        left: rect.left - containerRect.left,
-        width: rect.width,
-        opacity: 1,
-      });
-    }
-  };
-
-  const handleMouseLeave = () => {
-    setHoveredIdx(null);
-    setSpotlightPos((prev) => ({ ...prev, opacity: 0 }));
-  };
-
-  return {
-    hoveredIdx,
-    spotlightPos,
-    handleMouseEnter,
-    handleMouseLeave,
-  };
-};
