--- src/apps/web/src/features/navigation/components/MainMenu/index.ts
@@ -1,1 +1,0 @@
-export { MainMenu } from "./MainMenu";
