--- src/apps/web/src/features/navigation/components/MainMenu/types/MainMenuProps.ts
@@ -1,3 +1,0 @@
-export interface MainMenuProps {
-  mobile?: boolean;
-}
