--- src/apps/web/src/features/navigation/components/MainMenu/types/MainMenuSpotlight.ts
@@ -1,5 +1,0 @@
-export interface MainMenuSpotlight {
-  left: number;
-  width: number;
-  opacity: number;
-}
