+++ src/apps/web/src/features/navigation/components/MainMenuDesktop.tsx
@@ -1,14 +1,47 @@
-import { useRef } from "react";
+import { useRef, useState } from "react";
+import type { MouseEvent } from "react";
 
 import { cn } from "@repo/utils";
 
-import { NAV_LINKS } from "../constants/NavLinks";
-import { useMainMenuSpotlight } from "../hooks/useMainMenuSpotlight";
+interface SpotlightState {
+  left: number;
+  width: number;
+  opacity: number;
+}
+
+const NAV_LINKS = [
+  { href: "#features", label: "Vlastnosti" },
+  { href: "#motivation", label: "Proč my?" },
+  { href: "#pricing", label: "Ceník" },
+];
 
 export const MainMenuDesktop = () => {
   const containerRef = useRef<HTMLDivElement>(null);
-  const { hoveredIdx, spotlightPos, handleMouseEnter, handleMouseLeave } =
-    useMainMenuSpotlight(containerRef);
+  const [hoveredIdx, setHoveredIdx] = useState<number | null>(null);
+  const [spotlightPos, setSpotlightPos] = useState<SpotlightState>({
+    left: 0,
+    width: 0,
+    opacity: 0,
+  });
+
+  const handleMouseEnter = (idx: number, e: MouseEvent<HTMLAnchorElement>) => {
+    const rect = e.currentTarget.getBoundingClientRect();
+    const containerRect = containerRef.current?.getBoundingClientRect();
+
+    if (containerRect) {
+      setHoveredIdx(idx);
+      setSpotlightPos({
+        left: rect.left - containerRect.left,
+        width: rect.width,
+        opacity: 1,
+      });
+    }
+  };
+
+  const handleMouseLeave = () => {
+    setHoveredIdx(null);
+    setSpotlightPos((prev) => ({ ...prev, opacity: 0 }));
+  };
 
   return (
     <div
@@ -16,7 +49,6 @@ export const MainMenuDesktop = () => {
       className="relative flex items-center h-full"
       onMouseLeave={handleMouseLeave}
     >
-      {/* Liquid Spotlight Background */}
       <div
         className="absolute h-9 bg-white/10 rounded-full blur-[2px] transition-all duration-500 ease-out-quint pointer-events-none"
         style={{
@@ -26,7 +58,6 @@ export const MainMenuDesktop = () => {
           transform: `scale(${spotlightPos.opacity ? 1 : 0.8})`,
         }}
       >
-        {/* Glow core */}
         <div className="absolute inset-0 bg-primary/20 blur-md rounded-full animate-pulse" />
       </div>
 
@@ -44,7 +75,6 @@ export const MainMenuDesktop = () => {
         >
           {link.label}
 
-          {/* Animated dot indicator for active/hover state */}
           <div
             className={cn(
               "absolute -bottom-1 left-1/2 -translate-x-1/2 w-1 h-1 rounded-full bg-primary transition-all duration-500",
