+++ src/apps/web/src/features/navigation/index.ts
@@ -1,1 +1,2 @@
-export { MainMenu } from "./components/MainMenu";
+export { MainMenuDesktop } from "./components/MainMenuDesktop";
+export { MainMenuMobile } from "./components/MainMenuMobile";
