+++ src/apps/web/src/widgets/home/Pricing/Pricing.tsx
@@ -5,7 +5,30 @@ import { ButtonLink } from "@repo/components";
 import { Section } from "@web/shared/components/Section";
 
 import { PricingCard } from "./components/PricingCard";
-import { getPricingPlans } from "./data/pricingPlans";
+
+import type { TFunction } from "i18next";
+
+import type { PricingPlan } from "@web/entities/pricing";
+
+const getPricingPlans = (t: TFunction<"home">): PricingPlan[] => [
+  {
+    name: t("pricing.plans.free.name"),
+    price: 0,
+    features: t("pricing.plans.free.features", { returnObjects: true }),
+    cta: t("pricing.plans.free.cta"),
+    ctaLink: "/auth/register",
+  },
+  {
+    name: t("pricing.plans.premium.name"),
+    price: 49,
+    interval: t("pricing.plans.premium.interval"),
+    features: t("pricing.plans.premium.features", { returnObjects: true }),
+    cta: t("pricing.plans.premium.cta"),
+    ctaLink: "/auth/register",
+    badge: t("pricing.plans.premium.badge"),
+    highlighted: true,
+  },
+];
 
 export const Pricing = () => {
   const { t } = useTranslation("home");
