+++ src/apps/web/src/widgets/layout/Header/Header.tsx
@@ -1,15 +1,28 @@
+import { useEffect, useState } from "react";
+
+import { MainMenuDesktop } from "@web/features/navigation";
 import { Logo } from "@web/shared/components/Logo";
 
 import { ActionButtons } from "./components/ActionButtons";
-import { DesktopNav } from "./components/DesktopNav";
 import { MobileMenu } from "./components/MobileMenu";
 import { MobileToggle } from "./components/MobileToggle";
 import { useHeaderMobile } from "./hooks/useHeaderMobile";
-import { useHeaderScroll } from "./hooks/useHeaderScroll";
 import { headerVariants } from "./styles/headerVariants";
 
 export const Header = () => {
-  const { scrolled } = useHeaderScroll(20);
+  const [scrolled, setScrolled] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrolled(window.scrollY > 20);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    handleScroll();
+
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
   const { mobileOpen, toggleMobileMenu } = useHeaderMobile();
 
   const styles = headerVariants({ scrolled, mobileOpen });
@@ -21,8 +34,12 @@ export const Header = () => {
           <div className={styles.logoContainer()}>
             <Logo scrolled={scrolled} />
           </div>
+          <div className="hidden lg:flex items-center justify-center flex-1">
+            <nav className="flex items-center w-full justify-center">
+              <MainMenuDesktop />
+            </nav>
+          </div>
           <div className={styles.rightIconsContainer()}>
-            <DesktopNav />
             <ActionButtons scrolled={scrolled} />
             <MobileToggle
               mobileOpen={mobileOpen}
