+++ src/apps/web/src/widgets/layout/Header/components/MobileMenu.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 
 import { LanguageSelect } from "@web/features/locale";
-import { MainMenu } from "@web/features/navigation";
+import { MainMenuMobile } from "@web/features/navigation";
 import { ThemeSelect } from "@web/features/theme";
 
 interface MobileMenuProps {
@@ -17,7 +17,7 @@ export const MobileMenu = ({ styles }: MobileMenuProps) => {
         {/* 1. Navigation */}
         <nav className="flex-1">
           <div className="flex flex-col gap-1 [&_a]:flex [&_a]:items-center [&_a]:w-full [&_a]:py-4 [&_a]:px-4 [&_a]:text-[17px] [&_a]:font-medium [&_a]:text-slate-700 dark:[&_a]:text-slate-200 [&_.active]:text-blue-600 dark:[&_.active]:text-blue-400 [&_.active]:bg-blue-50 dark:[&_.active]:bg-blue-500/10 [&_.active]:rounded-2xl [&_a]:transition-all [&_a]:rounded-2xl [&_a:hover]:bg-slate-50 dark:[&_a:hover]:bg-white/5">
-            <MainMenu mobile />
+            <MainMenuMobile />
           </div>
         </nav>
