--- src/apps/web/src/widgets/layout/Header/hooks/useHeaderScroll.ts
@@ -1,19 +1,0 @@
-import { useEffect, useState } from "react";
-
-export const useHeaderScroll = (threshold = 20) => {
-  const [scrolled, setScrolled] = useState(false);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setScrolled(window.scrollY > threshold);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    handleScroll();
-
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, [threshold]);
-
-  return { scrolled };
-};
-
